Start migrating tests
